requirements.txt
fastapi
uvicorn[standard]

PyYAML
pillow
numpy
torch
timm
python-multipart
opencv-python